refactor: move status bar provider on top level
